requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.0
icrawler==0.6.7
idna==3.6
lxml==4.9.3
multidict==6.0.4
Pillow==10.1.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.1.0
yarl==1.9.4
